Fix error when checking volumes
If there is no match, grep will both error and return "0" causing the VOLUME_EXISTS var to be "0 0" due to the grep return and the echo 0. This causes the IF check to fail and error the entire script.

Simply make grep always succeed with a 'true' return value which still assigns VOLUME_EXISTS=0 when there are no matches